Refactor (BE-024): extract VO tests into their own test classes
