+++ src/test/kotlin/org/cescfe/bookpublishing/author/domain/model/AuthorTest.kt
@@ -1,9 +1,6 @@
 package org.cescfe.bookpublishing.author.domain.model
 
-import org.cescfe.bookpublishing.author.domain.exception.AuthorDomainException
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
-import java.util.UUID
 import kotlin.test.assertEquals
 
 class AuthorTest {
@@ -35,127 +32,5 @@ class AuthorTest {
         assertEquals(email, author.email)
         assertEquals(website, author.website)
         assertEquals(biography, author.biography)
-    }
-}
-
-class FullNameTest {
-    @Test
-    fun `should create full name with valid data`() {
-        // When
-        val fullName = FullName("John Ronald Reuel Tolkien")
-
-        // Then
-        assertEquals("John Ronald Reuel Tolkien", fullName.value)
-    }
-
-    @Test
-    fun `should throw AuthorValidationException when full name is blank`() {
-        // When & Then
-        val exception =
-            assertThrows<AuthorDomainException> {
-                FullName("   ")
-            }
-        assertEquals("Full name cannot be blank", exception.message)
-    }
-
-    @Test
-    fun `should throw AuthorValidationException when full name is too long`() {
-        // Given
-        val longName = "a".repeat(256)
-
-        // When & Then
-        val exception =
-            assertThrows<AuthorDomainException> {
-                FullName(longName)
-            }
-        assertEquals("Full name must be between 1 and 255 characters", exception.message)
-    }
-}
-
-class AuthorIdTest {
-    @Test
-    fun `should generate unique AuthorId`() {
-        // When
-        val id1 = AuthorId.generate()
-        val id2 = AuthorId.generate()
-
-        // Then
-        assertEquals(UUID::class.java, id1.value.javaClass)
-        assertEquals(UUID::class.java, id2.value.javaClass)
-        assert(id1 != id2)
-    }
-
-    @Test
-    fun `should create AuthorId from string`() {
-        // Given
-        val uuidString = "da420b0a-64aa-470d-991e-7fcb7a936229"
-
-        // When
-        val authorId = AuthorId.fromString(uuidString)
-
-        // Then
-        assertEquals(uuidString, authorId.value.toString())
-    }
-}
-
-class EmailTest {
-    @Test
-    fun `should create email with valid format`() {
-        // When
-        val email = Email("tolkien@example.com")
-
-        // Then
-        assertEquals("tolkien@example.com", email.value)
-    }
-
-    @Test
-    fun `should throw AuthorValidationException when email is blank`() {
-        // When & Then
-        assertThrows<AuthorDomainException> {
-            Email("   ")
-        }
-    }
-
-    @Test
-    fun `should throw AuthorValidationException when email does not contain @`() {
-        // When & Then
-        assertThrows<AuthorDomainException> {
-            Email("invalid-email")
-        }
-    }
-
-    @Test
-    fun `should throw AuthorValidationException when email format is invalid`() {
-        // When & Then
-        assertThrows<AuthorDomainException> {
-            Email("@example.com")
-        }
-    }
-}
-
-class WebsiteTest {
-    @Test
-    fun `should create website with valid URL`() {
-        // When
-        val website = Website("https://www.tolkiensociety.org")
-
-        // Then
-        assertEquals("https://www.tolkiensociety.org", website.value)
-    }
-
-    @Test
-    fun `should throw AuthorValidationException when website is blank`() {
-        // When & Then
-        assertThrows<AuthorDomainException> {
-            Website("   ")
-        }
-    }
-
-    @Test
-    fun `should throw AuthorValidationException when website does not start with http`() {
-        // When & Then
-        assertThrows<AuthorDomainException> {
-            Website("www.example.com")
-        }
     }
 }
